Restructure codebase around factory components. Clean typings.
