+++ src/lib/node/index.ts
@@ -1,6 +1,8 @@
+import { IProxiedNode } from './typings';
 import { NodeFactory } from './node';
 import { rerouteRPCMethodsHandler } from './rpc/rpcMethods';
-export type IProxiedNode = IAugmentedNode & IProxiedRpcMethods;
+
+export * from './typings'
 
 export const ProxiedNode = (endpoint: string): IProxiedNode => {
   const node = NodeFactory(endpoint);
