+++ src/lib/node/rpc/jsonUtils.ts
@@ -1,6 +1,7 @@
+import { ITransactionObject } from '../typings';
 import { randomBytes } from 'crypto';
 
-const generateId = (): String => randomBytes(16).toString('hex');
+export const generateId = (): String => randomBytes(16).toString('hex');
 
 export const generateTxObj = (tx: ITransactionObject) => ({
   id: generateId(),
